requirements.txt
Django==5.1.5
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
